chore(Engage): move sdk loader to different file
